requirements.txt
streamlit
Pillow
numpy


